updated iphone stuff
Hg-commit: db1e633e1c3db5b4ed9a4e54f4f04f87c0ffaf90

+++ xcode/examples/Classes/iphoneViewController.m
@@ -75,8 +75,8 @@ UITextView *globalTextView;
 */
 PetscErrorCode PetscVFPrintfiPhone(FILE *fd,const char *format,va_list Argp)
 {
-  int len;
-  char str[1024];
+  size_t len;
+  char   str[1024];
   
   PetscVSNPrintf(str,1024,format,&len,Argp);
   globalTextView.text = [NSString stringWithFormat:@"%@%s", globalTextView.text,str];
